+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * fix columns method
 * fix infinite loop when regex in formatters don't have g flag
 * fix parsing escape quotes
+* fix split equal to handle brackets when using without formatting
 
 ## 1.16.1
 ### Bugs